+++ code/autotask.cpp
@@ -136,7 +136,7 @@ void AutoTask::interfaceTimeout(const vector<gdioutput *> &windows) {
   lock = false;
 }
 
-void AutoTask::addSynchTime(DWORD tick) {
+void AutoTask::addSynchTime(uint32_t tick) {
   if (tick > 1000 * 60)
     return; // Ignore extreme times
 
@@ -146,7 +146,7 @@ void AutoTask::addSynchTime(DWORD tick) {
   synchQueue.push_back(tick);
 }
 
-DWORD AutoTask::getAvgSynchTime() {
+uint32_t AutoTask::getAvgSynchTime() {
 #ifdef NODELAY
   return 0;
 #else
@@ -165,10 +165,10 @@ DWORD AutoTask::getAvgSynchTime() {
 
 void AutoTask::synchronize(const vector<gdioutput *> &windows) {
   uint64_t tic = GetTickCount64();
-  DWORD avg = getAvgSynchTime();
+  uint32_t avg = getAvgSynchTime();
   //OutputDebugString(("AVG Update Time: " + itos(avg)).c_str());
   if (tic > lastSynchTime) {
-    DWORD since = tic - lastSynchTime;
+    uint32_t since = tic - lastSynchTime;
     if (since < avg * SYNC_FACTOR) {
       //OutputDebugString((" skipped: " + itos(since) + "\n").c_str());
       return;
@@ -191,7 +191,7 @@ void AutoTask::synchronize(const vector<gdioutput *> &windows) {
   if (synchronizeImpl(windows)) {
     uint64_t toc = GetTickCount64();
     if (toc > tic)
-      addSynchTime(toc-tic);
+      addSynchTime((uint32_t)(toc-tic));
 
     lastSynchTime = toc;
 #ifdef DEBUGPRINT
@@ -204,10 +204,10 @@ void AutoTask::synchronize(const vector<gdioutput *> &windows) {
 
 void AutoTask::advancePunchInformation(const vector<gdioutput *> &windows) {
   uint64_t tic = GetTickCount64();
-  DWORD avg = getAvgSynchTime();
+  uint32_t avg = getAvgSynchTime();
   //OutputDebugString(("Direct Update Time: " + itos(avg)).c_str());
   if (tic > lastSynchTime) {
-    DWORD since = tic-lastSynchTime;
+    uint32_t since = tic-lastSynchTime;
     if (since < avg * SYNC_FACTOR) {
       //OutputDebugString((" skipped: " + itos(since) + "\n").c_str());
       return;
@@ -216,8 +216,8 @@ void AutoTask::advancePunchInformation(const vector<gdioutput *> &windows) {
   else
     lastSynchTime = tic;
 
-  DWORD since = tic - lastTriedSynchTime;
-  if (since > DWORD(synchBaseTime*4)) { // Synchronize all instead.
+  uint32_t since = tic - lastTriedSynchTime;
+  if (since > uint32_t(synchBaseTime*4)) { // Synchronize all instead.
     synchronize(windows);
     return;
   }
@@ -225,7 +225,7 @@ void AutoTask::advancePunchInformation(const vector<gdioutput *> &windows) {
   if (advancePunchInformationImpl(windows)) {
     uint64_t toc = GetTickCount64();
     if (toc > tic)
-      addSynchTime(toc-tic);
+      addSynchTime((uint32_t)(toc-tic));
     lastSynchTime = toc;
 #ifdef DEBUGPRINT
     OutputDebugString((" direct update: " + itos(toc-tic) + "\n").c_str());
